pyproject.toml
[tool.poetry]
name = "knowledge-graph-for-rag-using-neo4j"
version = "0.1.0"
description = ""
authors = ["Homayoun.srp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
neo4j = "^5.23.1"
pypdf3 = "^1.0.6"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.0"
langchain-google-genai = "^2.0.5"
beautifulsoup4 = "^4.12.3"
sentence-transformers = "^3.0.0"
langchain-huggingface = "^0.1.0"
langchain-ollama = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
